# Metrics Generator Performance