--- a/modules/generator/processor/util/util.go
+++ b/modules/generator/processor/util/util.go
@@ -1,12 +1,16 @@
 package util
 
 import (
+	"slices"
+
 	semconv "go.opentelemetry.io/otel/semconv/v1.25.0"
 
 	v1_common "github.com/grafana/tempo/pkg/tempopb/common/v1"
 	v1_resource "github.com/grafana/tempo/pkg/tempopb/resource/v1"
 	v1 "github.com/grafana/tempo/pkg/tempopb/trace/v1"
 	tempo_util "github.com/grafana/tempo/pkg/util"
+
+	"github.com/prometheus/prometheus/util/strutil"
 )
 
 func FindServiceName(attributes []*v1_common.KeyValue) (string, bool) {
@@ -68,21 +72,30 @@ func GetJobValue(attributes []*v1_common.KeyValue) string {
 	return namespace + svName
 }
 
-func GetTargetInfoAttributesValues(attributes []*v1_common.KeyValue, exclude []string) ([]string, []string) {
+func GetTargetInfoAttributesValues(keys, values *[]string, attributes []*v1_common.KeyValue, exclude, intrinsicLabels []string) {
 	// TODO allocate with known length, or take new params for existing buffers
-	keys := make([]string, 0)
-	values := make([]string, 0)
+	*keys = (*keys)[:0]
+	*values = (*values)[:0]
 	for _, attrs := range attributes {
 		// ignoring job and instance
 		key := attrs.Key
-		value := tempo_util.StringifyAnyValue(attrs.Value)
 		if key != "service.name" && key != "service.namespace" && key != "service.instance.id" && !Contains(key, exclude) {
-			keys = append(keys, key)
-			values = append(values, value)
+			*keys = append(*keys, SanitizeLabelNameWithCollisions(key, intrinsicLabels))
+			value := tempo_util.StringifyAnyValue(attrs.Value)
+			*values = append(*values, value)
 		}
 	}
+}
 
-	return keys, values
+func SanitizeLabelNameWithCollisions(name string, dimensions []string) string {
+	sanitized := strutil.SanitizeLabelName(name)
+
+	// check if same label as intrinsics
+	if slices.Contains(dimensions, sanitized) {
+		return "__" + sanitized
+	}
+
+	return sanitized
 }
 
 func Contains(key string, list []string) bool {