<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

**What this PR does**: Generator improvements based on @mdisibio 's notes for things relating to target_info. 

```
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
          │ before-mg.out │            after-mg.out             │
          │    sec/op     │   sec/op     vs base                │
PushSpans     3.006m ± 3%   1.308m ± 1%  -56.49% (p=0.000 n=10)

          │ before-mg.out │            after-mg.out            │
          │  heap_in_use  │ heap_in_use  vs base               │
PushSpans     15.73M ± 0%   15.79M ± 0%  +0.36% (p=0.000 n=10)

          │ before-mg.out │             after-mg.out             │
          │     B/op      │     B/op      vs base                │
PushSpans    705.9Ki ± 0%   568.6Ki ± 0%  -19.44% (p=0.000 n=10)

          │ before-mg.out │            after-mg.out             │
          │   allocs/op   │  allocs/op   vs base                │
PushSpans    18.582k ± 0%   9.044k ± 0%  -51.33% (p=0.000 n=10)
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`