@ie-pham 2024-12-03
latest benchmarks


target info disabled
```
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
          │ before-tgdisabled-mg.out │      after-tgdisabled-mg.out       │
          │          sec/op          │   sec/op     vs base               │
PushSpans                2.102m ± 7%   1.969m ± 4%  -6.35% (p=0.000 n=10)

          │ before-tgdisabled-mg.out │      after-tgdisabled-mg.out       │
          │       heap_in_use        │ heap_in_use  vs base               │
PushSpans                15.86M ± 0%   16.00M ± 0%  +0.90% (p=0.000 n=10)

          │ before-tgdisabled-mg.out │    after-tgdisabled-mg.out     │
          │           B/op           │     B/op      vs base          │
PushSpans               540.2Ki ± 0%   540.2Ki ± 0%  ~ (p=0.063 n=10)

          │ before-tgdisabled-mg.out │    after-tgdisabled-mg.out    │
          │        allocs/op         │  allocs/op   vs base          │
PushSpans                15.36k ± 0%   15.36k ± 0%  ~ (p=0.839 n=10)
```

target info enabled
```
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
          │ before-tgenabled-mg.out │       after-tgenabled-mg.out        │
          │         sec/op          │   sec/op     vs base                │
PushSpans              7.455m ± 10%   3.263m ± 5%  -56.23% (p=0.000 n=10)

          │ before-tgenabled-mg.out │       after-tgenabled-mg.out       │
          │       heap_in_use       │ heap_in_use  vs base               │
PushSpans               16.36M ± 0%   16.15M ± 0%  -1.23% (p=0.000 n=10)

          │ before-tgenabled-mg.out │        after-tgenabled-mg.out        │
          │          B/op           │     B/op      vs base                │
PushSpans              1.724Mi ± 0%   1.386Mi ± 0%  -19.58% (p=0.000 n=10)

          │ before-tgenabled-mg.out │       after-tgenabled-mg.out        │
          │        allocs/op        │  allocs/op   vs base                │
PushSpans               46.44k ± 0%   22.60k ± 0%  -51.33% (p=0.000 n=10)
```